Rebrand to Plexamp for AAOS v1.0.0
- Renamed package from us.berkovitz.plexaaos to us.tiba.plexamp
- Updated LICENSE with dual copyright (original + modifications)
- Now uses JitPack for plexamp-for-aaos-api dependency
- Updated README with new branding and author info
- Updated authenticator and content provider authorities
- Version bump to 1.0.0

Based on original work by Joey Berkovitz (MIT License).
