+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
-Copyright (c) 2022 Joey Berkovitz
+Copyright (c) 2022 Joey Berkovitz (original work)
+Copyright (c) 2024-2025 Hamoudy Tiba (modifications and enhancements)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
