+++ automotive/src/main/java/us/tiba/plexamp/AuthenticatorService.kt
@@ -1,0 +1,19 @@
+package us.tiba.plexamp
+
+import android.app.Service
+import android.content.Intent
+import android.os.IBinder
+
+class AuthenticatorService : Service() {
+    private lateinit var authenticator: Authenticator
+
+    override fun onCreate() {
+        super.onCreate()
+
+        authenticator = Authenticator(this)
+    }
+
+    override fun onBind(intent: Intent): IBinder {
+        return authenticator.iBinder
+    }
+}
