+++ automotive/src/main/java/us/tiba/plexamp/extensions/JavaLangExt.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2018 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package us.tiba.plexamp.extensions
+
+import android.net.Uri
+
+/**
+ * Helper extension to convert a potentially null [String] to a [Uri] falling back to [Uri.EMPTY]
+ */
+fun String?.toUri(): Uri = this?.let { Uri.parse(it) } ?: Uri.EMPTY
