+++ automotive/src/main/res/xml/authenticator.xml
@@ -1,5 +1,5 @@
 <account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
-    android:accountType="us.berkovitz.plexaaos"
+    android:accountType="us.tiba.plexamp"
     android:icon="@drawable/ic_launcher_foreground"
     android:smallIcon="@drawable/ic_launcher_foreground"
     android:label="@string/app_name"
