+++ mobile/build.gradle
@@ -5,33 +5,23 @@ plugins {
 
 repositories {
     google()
-    mavenLocal()
     mavenCentral()
 }
 
 
 
 android {
-    signingConfigs {
-        release {
-            storeFile file('/home/joey/Documents/dev_store.jks')
-            storePassword 'password'
-            keyAlias 'key0'
-            keyPassword 'password'
-        }
-    }
-    namespace 'us.berkovitz.plexaaos'
+    namespace 'us.tiba.plexamp.mobile'
     compileSdk 33
 
     defaultConfig {
-        applicationId "us.berkovitz.plexaaos"
+        applicationId "us.tiba.plexamp.mobile"
         minSdk 28
         targetSdk 33
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        signingConfig signingConfigs.release
     }
 
     buildTypes {
