--- mobile/src/main/java/us/berkovitz/plexaaos/MainActivity.kt
@@ -1,11 +1,0 @@
-package us.berkovitz.plexaaos
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
